BB-41 Clean up Import and move Calu and Config in correct folder
