+++ src/main/java/com/butbetter/applicationservices/Calculator.java
@@ -1,6 +1,7 @@
 package com.butbetter.applicationservices;
 
 import com.butbetter.applicationservices.model.CalcApiResponse;
+import com.butbetter.applicationservices.service.CalculatorService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
