+++ src/main/java/com/butbetter/applicationservices/configuration/RestTemplateConfiguration.java
@@ -1,4 +1,4 @@
-package com.butbetter.applicationservices;
+package com.butbetter.applicationservices.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
